NFT: Separate core NFT logic from NFT card + handling of nft txmsg
