+++ node/lib/saito/ui/saito-nft/saito-nft.js
@@ -1,47 +1,181 @@
-const NftCreate = require('./overlays/create-overlay');
-const NftDisplay = require('./overlays/list-overlay');
-const NftDetails = require('./overlays/nft-overlay');
-const UIModTemplate = require('./../../../templates/uimodtemplate');
-
-/*
-  This is a container for all the independent overlays for displaying, creating, sending NFTs in Saito
-*/
-class SaitoNFT extends UIModTemplate {
-  constructor(app) {
-    super(app);
-
+class SaitoNft {
+  constructor(app, mod, tx = null, data = null, callback = null) {
     this.app = app;
-    this.name = 'SaitoNFT';
+    this.mod = mod;
 
-    //'saito-nft-create-render-request'
-    this.create = new NftCreate(app, this);
+    //
+    // nft details
+    //
+    this.id = data?.id;
+    this.tx_sig = data?.tx_sig;
+    this.slip1 = data?.slip1;
+    this.slip2 = data?.slip2;
+    this.slip3 = data?.slip3;
 
-    //'saito-nft-list-render-request'
-    this.display = new NftDisplay(app, this);
+    //
+    // tx details
+    //
+    this.tx = tx;
+    this.txmsg = null;
 
-    //'saito-nft-details-render-request'
-    this.details = new NftDetails(app, this);
+    this.amount = BigInt(0); // nolans
+    this.deposit = BigInt(0); // nolans
+    this.image = '';
+    this.text = '';
+
+    this.load_failed = false;
+
+    //
+    // UI helpers
+    //
+    this.uuid = null;
+    this.callback = callback;
+
+    //
+    // for auction
+    //
+    this.seller = '';
+    this.ask_price = BigInt(0);
+
+    this.buildNFTData();
   }
 
-  shouldAffixCallbackToModule() {
-    return 1;
+  async fetchTransaction(callback = null) {
+    if (!this.id) {
+      console.error('Unable to fetch NFT transaction (no nft id found)');
+      return;
+    }
+
+    this.app.storage.loadTransactions(
+      { field4: this.id },
+
+      async (txs) => {
+        if (txs?.length > 0) {
+          this.tx = txs[0];
+          this.buildNFTData();
+        } else {
+          //
+          // Try remote host (which **IS NOT** CURRENTLY INDEXING NFT TXS)
+          //
+          let peer = await this.app.network.getPeers();
+
+          this.app.storage.loadTransactions(
+            { field4: this.id },
+            (txs) => {
+              if (txs?.length > 0) {
+                this.tx = txs[0];
+
+                this.buildNFTData();
+                //
+                // save remotely fetched nft tx to local
+                //
+                this.app.storage.saveTransaction(this.tx, { field4: this.id }, 'localhost');
+              } else {
+                this.load_failed = true;
+              }
+            },
+            peer?.length ? peer[0] : null
+          );
+        }
+      },
+      'localhost'
+    );
   }
 
-  /***
-   *
-   * We can monitor all incoming txs on lite-blocks to see if they are nfts
-   *
-   */
-  async onConfirmation(blk, tx, conf) {
-    if (Number(conf) == 0) {
-      let txmsg = tx.returnMessage();
+  buildNFTData() {
+    let this_self = this;
+    if (!this.tx && !this.id) {
+      console.error('Insufficient data to make an nft!');
+      return;
+    }
 
-      if (txmsg.module == 'NFT') {
-        console.log(`UI Component SaitoNFT sees a NFT-marked transaction in Block ${blk.id}!!!`);
-        console.log(txmsg);
-      }
+    if (this.tx && this.id) {
+      console.log('Huzzah have a perfectly good nft!');
+      // already set
+      return;
+    }
+
+    if (this.tx) {
+      //
+      // tx is available we can extract slips & txmsg data (img/text)
+      //
+      this.extractNFTData();
+
+      this.slip1 = this.extractSlipObject(this.tx?.to[0] ?? null);
+      this.slip2 = this.extractSlipObject(this.tx?.to[1] ?? null);
+      this.slip3 = this.extractSlipObject(this.tx?.to[2] ?? null);
+    }
+
+    if (this.slip1?.amount) {
+      this.amount = BigInt(this.slip1.amount);
+    }
+
+    if (this.slip2?.amount) {
+      this.deposit = BigInt(this.slip2.amount);
+    }
+    this.uuid = this.slip1?.utxo_key;
+  }
+
+  //
+  // Extracts NFT image/text, tx_sig, txmsg data from a transaction
+  //
+  extractNFTData() {
+    if (!this.tx) {
+      return;
+    }
+
+    this.tx_sig = this.tx?.signature;
+    this.txmsg = this.tx.returnMessage();
+    this.id = this.computeNftIdFromTx(this.tx);
+
+    this.data = tx_msg?.data ?? {};
+
+    if (typeof this.data.image !== 'undefined') {
+      this.image = this.data.image;
+    }
+
+    if (typeof this.data.text !== 'undefined') {
+      this.text =
+        typeof this.data.text === 'object' && this.data.text !== null
+          ? JSON.stringify(this.data.text, null, 2)
+          : String(this.data.text);
     }
   }
+
+  extractSlipObject(slip) {
+    if (slip == null) return {};
+
+    const toStr = (v) => (typeof v === 'bigint' ? v.toString() : String(v));
+    const toNum = (v) => (typeof v === 'number' ? v : Number(v ?? 0));
+
+    return {
+      amount: toStr(slip.amount),
+      block_id: toStr(slip.blockId),
+      public_key: slip.publicKey,
+      slip_index: toNum(slip.index),
+      slip_type: toNum(slip.type),
+      tx_ordinal: toStr(slip.txOrdinal),
+      utxo_key: slip.utxoKey
+    };
+  }
+
+  //
+  // We need a way to get nft_id from NFT tx.
+  //
+  // If the NFT belongs to us we can simply get nft_id
+  // from storage (app.options.wallet.nfts[i].id). But in cases
+  // where NFT doesnt belong to us (e.g listed on assetstore) we
+  // need to compute nft_id based on the NFT tx we have.
+  //
+  // This situation isnt unqiue to assetstore, other mods will be
+  // creating NFT objects based on NFT tx so doesnt makes sense for this
+  // method below to be placed in assetstore.
+  //
+
+  //
+  // Ideal way would be to let rust comoute this by sending NFT tx
+  // to rust. For now temporarily JS is handling this.
+  //
 
   // Derive an NFT id from a tx
   computeNftIdFromTx(tx) {
@@ -80,8 +214,6 @@ class SaitoNFT extends UIModTemplate {
       .map((b) => b.toString(16).padStart(2, '0'))
       .join('');
   }
-
-  /* Helpers */
 
   hexToBytes(hex) {
     const clean = hex.startsWith('0x') ? hex.slice(2) : hex;
@@ -129,6 +261,31 @@ class SaitoNFT extends UIModTemplate {
     bytes.reverse();
     return new Uint8Array(bytes);
   }
+
+  async setPrice(saitoAmount) {
+    if (saitoAmount == null) throw new Error('setPrice: amount is required');
+    const saitoStr =
+      typeof saitoAmount === 'bigint' ? saitoAmount.toString() : String(saitoAmount).trim();
+    if (!saitoStr || isNaN(Number(saitoStr))) throw new Error('setPrice: invalid amount');
+    const nolan = await this.app.wallet.convertSaitoToNolan(saitoStr);
+    if (nolan == null) throw new Error('setPrice: conversion failed');
+    this.deposit = BigInt(nolan);
+    return this;
+  }
+
+  async getPrice() {
+    return await this.app.wallet.convertNolanToSaito(this.deposit);
+  }
+
+  async setSeller(public_key) {
+    if (public_key) {
+      this.seller = public_key;
+    }
+  }
+
+  async getSeller() {
+    return this.seller;
+  }
 }
 
-module.exports = SaitoNFT;
+module.exports = SaitoNft;
